requirements.txt
psycopg2-binary
PyYAML
python-dotenv